feat: create view for registration by generalizing review into preview component
