+++ packages/client/src/v2-events/features/events/actions/register/Register.tsx
@@ -10,25 +10,119 @@
  */
 
 import React from 'react'
-import { ROUTES } from '@client/v2-events/routes'
-import { EventDocument } from '@opencrvs/commons'
+import { defineMessages } from 'react-intl'
+import { useNavigate } from 'react-router-dom'
+import { v4 as uuid } from 'uuid'
 import { useTypedParams } from 'react-router-typesafe-routes/dom'
+import { ROUTES } from '@client/v2-events/routes'
 import { useEvents } from '@client/v2-events/features/events/useEvents/useEvents'
-import { ReviewSectionComponent } from '@client/v2-events/features/events/actions/declare/Review'
+import { Preview } from '@client/v2-events/features/events/components/Preview'
+import { useModal } from '@client/v2-events/hooks/useModal'
+import { useEventFormNavigation } from '@client/v2-events/features/events/useEventFormNavigation'
+import { useEventConfiguration } from '@client/v2-events/features/events/useEventConfiguration'
+import { useEventFormData } from '@client/v2-events/features/events/useEventFormData'
 
+const messages = defineMessages({
+  registerActionTitle: {
+    id: 'registerAction.title',
+    defaultMessage: 'Register member',
+    description: 'The title for register action'
+  },
+  registerActionDescription: {
+    id: 'registerAction.description',
+    defaultMessage:
+      'By clicking register, you confirm that the information entered is correct and the member can be registered.',
+    description: 'The description for register action'
+  },
+  registerActionDeclare: {
+    id: 'registerAction.Declare',
+    defaultMessage: 'Register',
+    description: 'The label for declare button of register action'
+  }
+})
+
+/**
+ *
+ * Preview of event to be registered.
+ */
 export const RegisterIndex = () => {
   const { eventId } = useTypedParams(ROUTES.V2.EVENTS.REGISTER.EVENT)
+  const events = useEvents()
+  const [modal, openModal] = useModal()
+  const navigate = useNavigate()
+  const { goToHome } = useEventFormNavigation()
+  const registerMutation = events.actions.declare()
 
-  const { getEvent } = useEvents()
-  const [event] = getEvent(eventId)
+  const [event] = events.getEvent(eventId)
 
-  if (!event) {
-    throw new Error('Event not found')
+  const { eventConfiguration: config } = useEventConfiguration(event.type)
+
+  if (!config) {
+    throw new Error('Event configuration not found with type: ' + event.type)
   }
 
-  return <Register event={event} />
-}
+  const { forms: formConfigs } = config.actions.filter(
+    (action) => action.type === 'REGISTER'
+  )[0]
 
-export const Register = ({ event }: { event: EventDocument }) => {
-  return <ReviewSectionComponent event={event} />
+  const form = useEventFormData((state) => state.formValues)
+
+  const handleEdit = async ({
+    pageId,
+    fieldId
+  }: {
+    pageId: string
+    fieldId?: string
+  }) => {
+    const confirmedEdit = await openModal<boolean | null>((close) => (
+      <Preview.EditModal close={close} />
+    ))
+
+    if (confirmedEdit) {
+      navigate(
+        ROUTES.V2.EVENTS.DECLARE.EVENT.PAGE.buildPath(
+          { pageId, eventId },
+          {
+            from: 'register'
+          },
+          fieldId
+        )
+      )
+    }
+    return
+  }
+
+  const handleRegistration = async () => {
+    const confirmedRegistration = await openModal<boolean | null>((close) => (
+      <Preview.ActionModal close={close} action="Register" />
+    ))
+    if (confirmedRegistration) {
+      registerMutation.mutate({
+        eventId: event.id,
+        data: form,
+        transactionId: `tmp-${uuid()}`
+      })
+
+      goToHome()
+    }
+  }
+
+  return (
+    <Preview.Body
+      formConfig={formConfigs[0]}
+      eventConfig={config}
+      onEdit={handleEdit}
+      form={form}
+    >
+      <Preview.Actions
+        onConfirm={handleRegistration}
+        messages={{
+          title: messages.registerActionTitle,
+          description: messages.registerActionDescription,
+          onConfirm: messages.registerActionDeclare
+        }}
+      />
+      {modal}
+    </Preview.Body>
+  )
 }
